Move ShareModal to src/Components/Share, delete dead Share.tsx
- ShareModal and ShareModalAutoPopup now live in Share/ alongside other share components
- Deleted Share.tsx and Share.css (dead code — Footer was the only consumer)
- Updated all imports in Footer.tsx, Results.tsx, and test files
